+++ src/components/Cart/CartMenu.tsx
@@ -1,33 +1,9 @@
-import {
-  Table,
-  TableBody,
-  TableCaption,
-  TableCell,
-  TableFooter,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import {
-  Sheet,
-  SheetClose,
-  SheetContent,
-  SheetHeader,
-  SheetFooter,
-  SheetTitle,
-  SheetTrigger,
-} from "@/components/ui/sheet";
-import { Button } from "@/components/ui/button";
+import { Sheet, SheetTrigger } from "@/components/ui/sheet";
 import {
   NavigationMenuLink,
   navigationMenuTriggerStyle,
 } from "@/components/ui/navigation-menu";
-// import { Cart } from "@/interface/Cart";
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const cartItems: any[] = [
-  // 장바구니에 들어있는 아이템 정보
-];
+import CartContents from "./CartContents";
 
 function CartMenu() {
   return (
@@ -37,43 +13,7 @@ function CartMenu() {
           장바구니
         </NavigationMenuLink>
       </SheetTrigger>
-      <SheetContent>
-        <SheetHeader>
-          <SheetTitle>장바구니</SheetTitle>
-        </SheetHeader>
-        <Table>
-          <TableCaption>장바구니에 추가된 상품 목록입니다.</TableCaption>
-          <TableHeader>
-            <TableRow>
-              <TableHead>상품명</TableHead>
-              <TableHead>가격</TableHead>
-              <TableHead>수량</TableHead>
-              <TableHead className="text-right">합계</TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            {cartItems.map((item) => (
-              <TableRow key={item.id}>
-                <TableCell>{item.name}</TableCell>
-                <TableCell>{item.price}</TableCell>
-                <TableCell>{item.quantity}</TableCell>
-                <TableCell className="text-right">{item.total}</TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-          <TableFooter>
-            <TableRow>
-              <TableCell colSpan={3}>총합</TableCell>
-              <TableCell className="text-right">{/*총합 계산*/}</TableCell>
-            </TableRow>
-          </TableFooter>
-        </Table>
-        <SheetFooter>
-          <SheetClose asChild>
-            <Button type="submit">결제하기</Button>
-          </SheetClose>
-        </SheetFooter>
-      </SheetContent>
+      <CartContents />
     </Sheet>
   );
 }
